Pull out the vibronic_utils information
